+++ .github/workflows/playwright.yml
@@ -20,6 +20,8 @@ jobs:
       run: npm ci
     - name: Start Supabase
       run: npx supabase start
+    - name: Run Jest tests
+      run: npx jest
     - name: Install Playwright Browsers
       run: npx playwright install --with-deps
     - name: Run Playwright tests
