+++ components/representatives/senators-table.tsx
@@ -1,35 +1,74 @@
-import { createClient } from "@/lib/supabase/server";
-import { SenatorsTableBody } from "@/components/representatives/senators-table-body";
+"use client";
+
+import { useEffect, useState, useTransition } from "react";
+import { useRouter } from "next/navigation";
+import { createClient } from "@/lib/supabase/client";
+import { Tables } from "@/lib/supabase/database.types";
+import { Badge } from "@/components/ui/badge";
+import { Button } from "@/components/ui/button";
+import {
+  Table,
+  TableBody,
+  TableCell,
+  TableHead,
+  TableHeader,
+  TableRow,
+} from "@/components/ui/table";
+
+type Representative = Tables<"representatives">;
 
 const PAGE_SIZE = 25;
 
-export async function SenatorsTable({
-  searchParams,
-}: {
-  searchParams: Promise<{ page?: string }>;
-}) {
-  const { page: pageParam } = await searchParams;
-  const page = Math.max(0, parseInt(pageParam ?? "0", 10) || 0);
-  const supabase = await createClient();
-  const from = page * PAGE_SIZE;
-  const to = from + PAGE_SIZE - 1;
+const partyColor: Record<string, string> = {
+  Democrat: "bg-blue-100 text-blue-800 dark:bg-blue-900 dark:text-blue-200",
+  Republican: "bg-red-100 text-red-800 dark:bg-red-900 dark:text-red-200",
+  Independent:
+    "bg-purple-100 text-purple-800 dark:bg-purple-900 dark:text-purple-200",
+};
 
-  const { count } = await supabase
-    .from("representatives")
-    .select("id", { count: "exact", head: true })
-    .eq("chamber", "sen")
-    .eq("in_office", true);
+export function SenatorsTable() {
+  const router = useRouter();
+  const [page, setPage] = useState(0);
+  const [senators, setSenators] = useState<Representative[]>([]);
+  const [totalPages, setTotalPages] = useState(0);
+  const [error, setError] = useState(false);
+  const [isPending, startTransition] = useTransition();
 
-  const totalPages = Math.ceil((count ?? 0) / PAGE_SIZE);
+  useEffect(() => {
+    startTransition(async () => {
+      setError(false);
+      const supabase = createClient();
+      const from = page * PAGE_SIZE;
+      const to = from + PAGE_SIZE - 1;
 
-  const { data: senators, error } = await supabase
-    .from("representatives")
-    .select()
-    .eq("chamber", "sen")
-    .eq("in_office", true)
-    .order("state")
-    .order("last_name")
-    .range(from, to);
+      const { count } = await supabase
+        .from("representatives")
+        .select("id", { count: "exact", head: true })
+        .eq("chamber", "sen")
+        .eq("in_office", true);
+
+      const { data, error: queryError } = await supabase
+        .from("representatives")
+        .select()
+        .eq("chamber", "sen")
+        .eq("in_office", true)
+        .order("state")
+        .order("last_name")
+        .range(from, to);
+
+      if (queryError) {
+        setError(true);
+      } else {
+        setSenators(data ?? []);
+        setTotalPages(Math.ceil((count ?? 0) / PAGE_SIZE));
+      }
+    });
+  }, [page]);
+
+  if (isPending) {
+    // TODO(lfm): Could be nice to change this to a better table loading state.
+    return <p className="text-muted-foreground">Loading…</p>;
+  }
 
   if (error) {
     return <p className="text-destructive">Failed to load senators.</p>;
@@ -44,10 +83,68 @@ export async function SenatorsTable({
   }
 
   return (
-    <SenatorsTableBody
-      senators={senators}
-      page={page}
-      totalPages={totalPages}
-    />
+    <div className="max-w-[50%]">
+      <div className="overflow-hidden rounded-md border">
+        <Table>
+          <TableHeader>
+            <TableRow className="bg-primary hover:bg-primary">
+              <TableHead className="text-primary-foreground">Name</TableHead>
+              <TableHead className="text-primary-foreground">State</TableHead>
+              <TableHead className="text-primary-foreground">Party</TableHead>
+            </TableRow>
+          </TableHeader>
+          <TableBody>
+            {senators.map((senator) => (
+              <TableRow
+                key={senator.id}
+                className="cursor-pointer hover:underline"
+                onClick={() =>
+                  router.push(`/representatives/${senator.bioguide_id}`)
+                }
+              >
+                <TableCell className="font-medium">
+                  {senator.official_full_name ??
+                    `${senator.first_name} ${senator.last_name}`}
+                </TableCell>
+                <TableCell>{senator.state}</TableCell>
+                <TableCell>
+                  <Badge
+                    variant="outline"
+                    className={partyColor[senator.party] ?? ""}
+                  >
+                    {senator.party[0]}
+                  </Badge>
+                </TableCell>
+              </TableRow>
+            ))}
+          </TableBody>
+        </Table>
+      </div>
+      {totalPages > 1 && (
+        <div className="flex items-center justify-between pt-4">
+          <p className="text-sm text-muted-foreground">
+            Page {page + 1} of {totalPages}
+          </p>
+          <div className="flex gap-2">
+            <Button
+              variant="outline"
+              size="sm"
+              onClick={() => setPage((p) => p - 1)}
+              disabled={page === 0}
+            >
+              Previous
+            </Button>
+            <Button
+              variant="outline"
+              size="sm"
+              onClick={() => setPage((p) => p + 1)}
+              disabled={page >= totalPages - 1}
+            >
+              Next
+            </Button>
+          </div>
+        </div>
+      )}
+    </div>
   );
 }
